+++ src/unix/async_impl/smol_impl.rs
@@ -3,7 +3,7 @@ use std::os::unix::fs::MetadataExt;
 use std::os::unix::io::AsRawFd;
 
 lock_impl!(File);
-allocate!(File);
+allocate!(File, smol);
 allocate_size!(File);
 
 test_mod! {